requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==4.0.3
attrs==25.4.0
bcrypt==5.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
contourpy==1.3.0
cryptography==46.0.4
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
ecdsa==0.19.1
exceptiongroup==1.3.1
fastapi==0.128.1
ffmpy==1.0.0
filelock==3.19.1
fonttools==4.60.2
frozenlist==1.8.0
fsspec==2025.10.0
gradio==4.44.1
gradio_client==1.3.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.19.4
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.7
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.83
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langsmith==0.4.37
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.2
matplotlib==3.9.4
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.16.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
openai==2.16.0
openai-whisper==20250625
orjson==3.11.5
packaging==26.0
pandas==2.3.3
passlib==1.7.4
pillow==10.4.0
propcache==0.4.1
pyasn1==0.6.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.2
rpds-py==0.27.1
rsa==4.9.1
ruff==0.15.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.5
soundfile==0.13.1
SQLAlchemy==2.0.46
starlette==0.49.3
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
tomlkit==0.12.0
torch==2.8.0
tqdm==4.67.3
typer==0.21.1
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.39.0
websockets==11.0.3
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
